Only download {babylon,hand}.js if they don't already exist.

+++ scripts/install-pages.sh
@@ -10,8 +10,13 @@ output=$outputbase/GrappleMap
 
 mkdir -p $output/{gifframes,composer,search}
 
-wget --no-verbose -O $output/babylon.js "https://raw.githubusercontent.com/BabylonJS/Babylon.js/master/dist/preview%20release/babylon.js"
-wget --no-verbose -O $output/hand.js "https://raw.githubusercontent.com/deltakosh/handjs/master/bin/hand.min.js"
+if [ ! -f $output/babylon.js ]; then
+	wget --no-verbose -O $output/babylon.js "https://raw.githubusercontent.com/BabylonJS/Babylon.js/master/dist/preview%20release/babylon.js"
+fi
+
+if [ ! -f $output/hand.js ]; then
+	wget --no-verbose -O $output/hand.js "https://raw.githubusercontent.com/deltakosh/handjs/master/bin/hand.min.js"
+fi
 
 cp src/gm.js $output/
 cp src/composer.html $output/composer/index.html
